refactored code to remove extra `section/header/title` and `section/header/description` subcomponents + contextual improvements to showcase
