+++ packages/components/package.json
@@ -256,9 +256,7 @@
       "./components/hds/form/radio/field.js": "./dist/_app_/components/hds/form/radio/field.js",
       "./components/hds/form/radio/group.js": "./dist/_app_/components/hds/form/radio/group.js",
       "./components/hds/form/section/field-group/index.js": "./dist/_app_/components/hds/form/section/field-group/index.js",
-      "./components/hds/form/section/header/description.js": "./dist/_app_/components/hds/form/section/header/description.js",
-      "./components/hds/form/section/header/index.js": "./dist/_app_/components/hds/form/section/header/index.js",
-      "./components/hds/form/section/header/title.js": "./dist/_app_/components/hds/form/section/header/title.js",
+      "./components/hds/form/section/header.js": "./dist/_app_/components/hds/form/section/header.js",
       "./components/hds/form/section/index.js": "./dist/_app_/components/hds/form/section/index.js",
       "./components/hds/form/select/base.js": "./dist/_app_/components/hds/form/select/base.js",
       "./components/hds/form/select/field.js": "./dist/_app_/components/hds/form/select/field.js",
