+++ packages/components/src/components/hds/form/index.hbs
@@ -12,8 +12,8 @@
         HeaderDescription=(component "hds/form/header/description")
         Section=(component "hds/form/section")
         SectionHeader=(component "hds/form/section/header")
-        SectionHeaderTitle=(component "hds/form/section/header/title")
-        SectionHeaderDescription=(component "hds/form/section/header/description")
+        SectionHeaderTitle=(component "hds/form/header/title" size="300")
+        SectionHeaderDescription=(component "hds/form/header/description")
         SectionFieldGroup=(component "hds/form/section/field-group")
         Separator=(component "hds/form/separator")
       )
@@ -28,8 +28,8 @@
         HeaderDescription=(component "hds/form/header/description")
         Section=(component "hds/form/section")
         SectionHeader=(component "hds/form/section/header")
-        SectionHeaderTitle=(component "hds/form/section/header/title")
-        SectionHeaderDescription=(component "hds/form/section/header/description")
+        SectionHeaderTitle=(component "hds/form/header/title" size="300")
+        SectionHeaderDescription=(component "hds/form/header/description")
         SectionFieldGroup=(component "hds/form/section/field-group")
         Separator=(component "hds/form/separator")
       )
