+++ packages/components/src/components/hds/form/section/header.hbs
@@ -5,7 +5,7 @@
 <div class="hds-form__section-header" ...attributes>
   {{yield
     (hash
-      Title=(component "hds/form/section/header/title") Description=(component "hds/form/section/header/description")
+      Title=(component "hds/form/header/title" size="300") Description=(component "hds/form/header/description")
     )
   }}
 </div>
