--- packages/components/src/components/hds/form/section/header/description.hbs
@@ -1,8 +1,0 @@
-{{!
-  Copyright (c) HashiCorp, Inc.
-  SPDX-License-Identifier: MPL-2.0
-}}
-
-<Hds::Text::Body class="hds-form__section-header-description" @tag="p" @size="300" @color="primary" ...attributes>
-  {{yield}}
-</Hds::Text::Body>
