--- packages/components/src/components/hds/form/section/header/title.hbs
@@ -1,14 +1,0 @@
-{{!
-  Copyright (c) HashiCorp, Inc.
-  SPDX-License-Identifier: MPL-2.0
-}}
-
-<Hds::Text::Display
-  class="hds-form__section-header-title"
-  @tag={{this.tag}}
-  @size={{this.size}}
-  @color="strong"
-  ...attributes
->
-  {{yield}}
-</Hds::Text::Display>
