+++ packages/components/src/components/hds/form/section/index.hbs
@@ -7,8 +7,8 @@
   {{yield
     (hash
       Header=(component "hds/form/section/header")
-      HeaderTitle=(component "hds/form/section/header/title")
-      HeaderDescription=(component "hds/form/section/header/description")
+      HeaderTitle=(component "hds/form/header/title" size="300")
+      HeaderDescription=(component "hds/form/header/description")
       SectionFieldGroup=(component "hds/form/section/field-group")
       Section=(component "hds/form/section")
     )
