--- showcase/tests/integration/components/hds/form/section/header/description/index-test.js
@@ -1,25 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-import { module, test } from 'qunit';
-import { setupRenderingTest } from 'showcase/tests/helpers';
-import { render } from '@ember/test-helpers';
-import { hbs } from 'ember-cli-htmlbars';
-
-module(
-  'Integration | Component | hds/form/section/header/description/index',
-  function (hooks) {
-    setupRenderingTest(hooks);
-
-    test('it should render the component with a CSS class that matches the component name', async function (assert) {
-      await render(
-        hbs`<Hds::Form::Section::Header::Description id="test-form-description" />`,
-      );
-      assert
-        .dom('#test-form-description')
-        .hasClass('hds-form__section-header-description');
-    });
-  },
-);
